Refactor - Break up API file
